+++ scripts/sara_rest_cli.py
@@ -26,12 +26,14 @@ def make_request(method, url, email):
         response = requests.get(url, headers=signed_header_value)
     elif method == "POST":
         response = requests.post(url, headers=signed_header_value)
+    elif method == "DELETE":
+        response = requests.delete(url, headers=signed_header_value)
     else:
         raise ValueError("Unsupported method")
     return response
 
 
-def main(email, org, project, method, stage):
+def main(email, org, project, method, stage, data):
     URL = stage_url[stage]
     endpoints = {
         "status": f"{URL}/api/user_project/{org}/{project}/status",
@@ -47,9 +49,19 @@ def main(email, org, project, method, stage):
         "aispec_gen": f"{URL}/api/user_project/{org}/{project}/data/aispec/generator",
         "blueprint_gen": f"{URL}/api/user_project/{org}/{project}/data/blueprint/generator",
 
+        "projectsource_status": f"{URL}/api/user_project/{org}/{project}/data/projectsource/status",
+        "aispec_status": f"{URL}/api/user_project/{org}/{project}/data/aispec/status",
+        "blueprint_status": f"{URL}/api/user_project/{org}/{project}/data/blueprint/status",
+
         "create_blueprint": f"{URL}/api/user_project/{org}/{project}/data/blueprint/generator/start",
         "create_aispec": f"{URL}/api/user_project/{org}/{project}/data/aispec/generator/start",
-        "create_projectsource": f"{URL}/api/user_project/{org}/{project}/data/projectsource/generator/start"
+        "create_projectsource": f"{URL}/api/user_project/{org}/{project}/data/projectsource/generator/start",
+
+        "aifiles": f"{URL}/api/user/{org}/connectors/openai/files",
+        "aifiles_groom": f"{URL}/api/user/{org}/connectors/openai/files/?groom",
+        "aifile_delete": f"{URL}/api/user/{org}/connectors/openai/files/{data}",
+
+        "assistants": f"{URL}/api/user/{org}/connectors/openai/assistants",
     }
 
     if method not in endpoints:
@@ -58,14 +70,15 @@ def main(email, org, project, method, stage):
 
     url = endpoints[method]
     # if method starts with "create_" or is "discovery", then it's a POST request
-    response = make_request("POST" if method.startswith("create_") or method == "discovery" else "GET", url, email)
-
-    # Output the response
-    print(response.status_code)
+    verb = "POST" if method.startswith("create_") or method == "discovery" else "DELETE" if (method == "aifiles_groom" or method == "aifile_delete") else "GET"
+    print(f"Requesting {verb} {url}")
+    response = make_request(verb, url, email)
 
     if (response.status_code != 200):
         print(f"Failed({response.status_code}): {response.text}")
     else:
+        print(f"Success({response.status_code})\n")
+
         if response.headers.get('content-type') == 'application/json':
             print(response.json() if 'body' not in response.json() else json.loads(response.json()['body']))
         else:
@@ -76,9 +89,38 @@ if __name__ == "__main__":
     parser = argparse.ArgumentParser(description="CLI utility for managing user projects.")
     parser.add_argument("--email", required=True, help="The user's email address")
     parser.add_argument("--org", default="polyverse-appsec", help="The organization name (default: polyverse-appsec)")
-    parser.add_argument("--project", required=True, help="The project name")
-    parser.add_argument("--method", default="status", choices=['status', 'discovery', 'data_references', 'projectsource', 'aispec', 'blueprint', 'blueprint_gen', 'projectsource_gen', 'aispec_gen'], help="The method to run")
+    parser.add_argument("--project", required=False, help="The project name")
+    parser.add_argument("--method", default="status",
+                        choices=['status',
+
+                                 'discovery',
+                                 'data_references',
+
+                                 'projectsource',
+                                 'aispec',
+                                 'blueprint',
+
+                                 'blueprint_gen',
+                                 'projectsource_gen',
+                                 'aispec_gen',
+
+                                 'blueprint_status',
+                                 'projectsource_status',
+                                 'aispec_status',
+
+                                 'aifiles',
+                                 'aifiles_groom',
+                                 'aifile_delete',
+
+                                 'assistants'
+                                 ], help="The method to run")
     parser.add_argument("--stage", default="local", choices=['local', 'dev', 'test', 'prod'], help="The Service to target (default: local)")
+    parser.add_argument("--data", default=None, help="Data to pass to the method")
 
     args = parser.parse_args()
-    main(args.email, args.org, args.project, args.method, args.stage)
+
+    if (args.project is None and args.method not in ["status", "data_references", "aifiles", "aifiles_groom", "aifile_delete", "assistants"]):
+        parser.error("The --project argument is required for the method"
+                     f" {args.method}.")
+
+    main(args.email, args.org, args.project, args.method, args.stage, args.data)